pyproject.toml
[project]
name = "on_the_fly_resume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "ipykernel>=7.1.0",
    "jinja2>=3.1.6",
    "openai>=2.15.0",
    "pdflatex>=0.1.3",
    "poethepoet>=0.24.0",
    "pydantic>=2.12.5",
    "pyyaml>=6.0.3",
]

[tool.uv]
package = true

[build-system]
requires = ["setuptools>=61.0","wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "on_the_fly_resume" }

[tool.setuptools.packages.find]
where = ["on_the_fly_resume"]

[tool.poe.tasks]
template = "python -m on_the_fly_resume.__main__"
render = "pdflatex -output-directory=${RENDER_DIR} ./output/${OUTPUT}"
generate = { sequence = ["template", "render"], env = { TEMPLATE = "template.jinja", DATA = "resume_content.yaml", OUTPUT = "output.tex", RENDER_DIR = "./render" } }
